+++ src/main/java/org/folio/rest/impl/InstanceBatchSyncUnsafeApi.java
@@ -8,11 +8,9 @@ import io.vertx.core.Handler;
 import java.util.Map;
 import javax.ws.rs.core.Response;
 import org.folio.rest.annotations.Validate;
-import org.folio.rest.jaxrs.model.Instances;
 import org.folio.rest.jaxrs.model.InstancesPost;
 import org.folio.rest.jaxrs.resource.InstanceStorageBatchSynchronousUnsafe;
 import org.folio.services.instance.InstanceService;
-import org.folio.utils.ObjectConverterUtils;
 
 public class InstanceBatchSyncUnsafeApi implements InstanceStorageBatchSynchronousUnsafe {
   @Validate
@@ -21,10 +19,8 @@ public class InstanceBatchSyncUnsafeApi implements InstanceStorageBatchSynchrono
                                                         Handler<AsyncResult<Response>> asyncResultHandler,
                                                         Context vertxContext) {
 
-    var instances = ObjectConverterUtils.convertObject(entity, Instances.class);
-
     new InstanceService(vertxContext, okapiHeaders)
-      .createInstances(instances.getInstances(), true, false, true)
+      .createInstances(entity.getInstances(), true, false, true)
       .onFailure(handleFailure(asyncResultHandler))
       .onComplete(asyncResultHandler);
   }