+++ src/main/java/org/folio/services/migration/async/AsyncMigrationJobService.java
@@ -32,8 +32,7 @@ import org.folio.rest.persist.PostgresClientFuturized;
 public final class AsyncMigrationJobService {
   private static final List<AsyncMigrationJobRunner> MIGRATION_JOB_RUNNERS = List
     .of(new ShelvingOrderMigrationJobRunner(),
-      new SubjectSeriesMigrationJobRunner(),
-      new PublicationPeriodMigrationJobRunner());
+      new SubjectSeriesMigrationJobRunner());
   private static final List<AsyncMigrationJob.JobStatus> ACCEPTABLE_STATUSES = List
     .of(AsyncMigrationJob.JobStatus.IN_PROGRESS, IDS_PUBLISHED);